Delete the old Rotkehlchen Server and replace with simple version
